requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.1.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
future==0.18.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
idna==2.10
jsonschema==3.2.0
Logbook==1.5.3
matrix-nio==0.15.1
multidict==4.7.6
peewee==3.13.3
pycparser==2.20
pycryptodome==3.9.8
pyrsistent==0.16.0
python-olm==3.1.3
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
unpaddedbase64==1.1.0
urllib3==1.25.10
yarl==1.5.1
